refactor: recreate state with createStore()
