+++ src/AsideContainer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Aside from './Aside';
+
+function AsideContainer(props) {
+
+  return (
+    <Aside linkData = { props.linkData } dialogsData = {props.dialogsData}/>
+  );
+}
+
+export default AsideContainer;
