+++ src/components/data/sidebar-reducer.js
@@ -7,9 +7,9 @@ let initialState = [
     // { id: 6, link: "/friends", linkName: "friends"},
 ]
 
-const sidebarReduser = (state = initialState, action) => {
+const sidebarReducer = (state = initialState, action) => {
 
     return state;
 }
 
-export default sidebarReduser;
+export default sidebarReducer;
